FEATURE: Build static components
