+++ src/shoppingList.js
@@ -23,8 +23,25 @@ ShoppingListRow.propTypes = {
 }
 
 class ShoppingListTable extends Component {
+    constructor(props){
+        super(props);
+        this.state = {
+            shoppingLists: []
+        };
+    }
+
+    componentDidMount(){
+        const shoppingLists = [
+            {"name": "Food", "id": "1", "due_date": "2017-07-09"},
+            {"name": "School items", "id": "2", "due_date": "2017-08-09"},
+            {"name": "Hardware", "id": "3", "due_date": "2017-07-29"}
+          ]        
+        this.setState({
+            shoppingLists,
+        });
+    }
     render(){
-        const rows = this.props.shoppingLists.map((shoppingList) => {
+        const rows = this.state.shoppingLists.map((shoppingList) => {
             return(
                 <ShoppingListRow
                 shoppingList={shoppingList}
